Bug fix - primary user of the chat cannot delete himself from the chat (it is supposed that the whole chat should be deleted for that purpose). Tiny adjustments.

+++ src/assets/js/components/DeleteUsersList/index.ts
@@ -10,34 +10,40 @@ const globalStateInstance = new GlobalState();
 
 export default class DeleteUsersList extends Block {
     constructor(props: IDeleteUsersListProps) {
+        const currentUser = globalStateInstance.getProp("currentUser");
         super("div", {
             ...props,
-            users: props.users.map((user: IUser) => ({
-                login: user.login,
-                deleteButton: new Button("button", {
-                    text: "Delete",
-                    attributes: {
-                        class: "button button_danger button_wide"
-                    },
-                    eventListeners: [
-                        ["click", () => {
-                            const selectedChat: IExistingChat = globalStateInstance.getProp("selectedChat");
-                            const userRemovalData = {
-                                users: [user.id],
-                                chatId: selectedChat.id
-                            };
-                            chatsAPI.deleteUsers(userRemovalData)
-                                .then(() => {
-                                    console.log(`[SUCCESS] User ${user.login} has been successfully removed from chat ${selectedChat.title}`);
-                                    const oldProps = this.props;
-                                    const newUsers = oldProps.users.filter((renderedUser: IDeleteUserProps) => renderedUser.login !== user.login);
-                                    this.setProps({users: newUsers});
-                                })
-                                .catch((error: XMLHttpRequest) => console.log(`[ERROR] Could not delete user ${user.login} from chat ${selectedChat.title}`, JSON.parse(error.response)))
-                        }]
-                    ]
-                })
-            }))
+            users: props.users.reduce((acc: IDeleteUserProps[], user: IUser) => {
+                if (user.id !== currentUser.id) {
+                    acc.push({
+                        login: user.login,
+                        deleteButton: new Button("button", {
+                            text: "Delete",
+                            attributes: {
+                                class: "button button_danger button_wide"
+                            },
+                            eventListeners: [
+                                ["click", () => {
+                                    const selectedChat: IExistingChat = globalStateInstance.getProp("selectedChat");
+                                    const userRemovalData = {
+                                        users: [user.id],
+                                        chatId: selectedChat.id
+                                    };
+                                    chatsAPI.deleteUsers(userRemovalData)
+                                        .then(() => {
+                                            console.log(`[SUCCESS] User ${user.login} has been successfully removed from chat ${selectedChat.title}`);
+                                            const oldProps = this.props;
+                                            const newUsers = oldProps.users.filter((renderedUser: IDeleteUserProps) => renderedUser.login !== user.login);
+                                            this.setProps({users: newUsers});
+                                        })
+                                        .catch((error: XMLHttpRequest) => console.log(`[ERROR] Could not delete user ${user.login} from chat ${selectedChat.title}`, JSON.parse(error.response)))
+                                }]
+                            ]
+                        })
+                    });
+                }
+                return acc;
+            }, [])
         });
     }
